Implement logic to remove current courses in timetable & add new courses if a valid timetable combination is found
