+++ backend/api/timetable-model.js
@@ -43,4 +43,41 @@ module.exports = {
       throw new Error();
     }
   },
+
+  async addTimetableCourse(term, position, courseId, timetableId, userId) {
+    await prisma.user.update({
+      where: {
+        id: userId,
+      },
+      data: {
+        timetables: {
+          update: {
+            where: {
+              id: timetableId,
+            },
+            data: {
+              courses: {
+                upsert: {
+                  where: {
+                    courseId_timetableId: { courseId, timetableId },
+                  },
+                  create: {
+                    term,
+                    position,
+                    course: {
+                      connect: { id: courseId },
+                    },
+                  },
+                  update: {
+                    term,
+                    position,
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    });
+  },
 };
